Enabling wsign-compare check and fixing 32bit failures

+++ utils/s2n_mem.c
@@ -51,7 +51,7 @@ static int s2n_mem_init_impl(void)
     POSIX_ENSURE_GT(sysconf_rc, 0);
 
     /* page_size must be a valid uint32 */
-    POSIX_ENSURE_LTE(sysconf_rc, UINT32_MAX);
+    POSIX_ENSURE_LTE((uint32_t) sysconf_rc, UINT32_MAX);
 
     page_size = (uint32_t) sysconf_rc;
